[Silver V] Title: LCM, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/5347. LCM/LCM.cc
@@ -1,13 +1,13 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-long long gcd(long long a, long long b){
+int gcd(int a, int b){
     if(a==0) return b;
     return gcd(b%a,a);
 }
 
-long long lcm(long long a, long long b){
-    return a/gcd(a,b)*b;
+long long lcm(int a, int b){
+    return (long long)a/gcd(a,b)*b;
 }
 
 int main(){
@@ -16,7 +16,7 @@ int main(){
     int t;
     cin>>t;
     while(t--){
-        long long a, b;
+        int a, b;
         cin>>a>>b;
         cout<<lcm(a,b)<<'\n';
     }
